app/requirements.txt
Django
djangorestframework
pyscopg2
wsgi